refactor(web): Phase 1 — lift settings into 3-tier (Infra → Service → API)
First domain migrated under `docs/web-architecture.md`'s API → Services →
Infra layout. Settings was chosen as the pilot because the surface is
tiny (load/merge/save a JSON document plus three derived helpers), which
lets the tier boundaries be the point of the change rather than fighting
business-logic complexity.

- Infra: `server/infra/db/queries/settings.ts` — extract the
  `~/.band/settings.json` file I/O (atomic temp-file write + merge on
  load) plus the `bandHome()` helper and all settings types out of
  `lib/state.ts` into a stateless `SettingsQueries` class.
- Service: `server/services/settings-service.ts` — own all business
  logic that used to be inline in `lib/state.ts`: agent-definition
  fallback chain (`resolveAgent` is exposed as a static so the back-compat
  shim can delegate without a re-load), `worktreesDir()` default, and
  token bootstrap. Module-level `settingsService` singleton consumed by
  both the API tier and the legacy `lib/state` wrappers.
- API: `server/api/settings/router.ts` — thin tRPC sub-router that
  validates input with Zod and delegates to the service. `update()`
  takes `Partial<Settings>` to make the merge semantics explicit.
- API merge point: `server/api/router.ts` switches from a re-export
  shim to a real `t.mergeRouters(legacyAppRouter, t.router({ settings }))`
  composition. The new shared tRPC builder lives at
  `server/api/trpc.ts` so the merged routers share an instance (tRPC's
  `mergeRouters` requires it).
- Legacy router shrunk: removed the inline settings declaration from
  `trpc/router.ts` (now imports `t`/`publicProcedure` from the shared
  builder). All other domains stay where they are — they'll move in
  their own phases.
- Back-compat wrappers in `lib/state.ts` preserve the `loadSettings` /
  `saveSettings` / `getAgentDefinition` / `getOrCreateToken` /
  `worktreesDir` import surface so the rest of the codebase (chat-manager,
  agent-pool, setup, …) keeps compiling. Subsequent phases will rewrite
  those callers and delete the shim.
- `server/infra/db/connection.ts` now imports `bandHome` from its
  sibling queries module instead of the legacy `lib/state` re-export.
- Delete the `services/.gitkeep` placeholder now that a real service
  file lives in the directory.

Wire shape is unchanged: `/trpc/settings.get` and `/trpc/settings.update`
still surface in the generated OpenAPI document and the existing
integration tests (`apps/web/tests/trpc.test.ts`) pass end-to-end against
the real server through the new 3-tier stack.

Closes #312
